pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ml-detection" # fill this
version = "0.0.0"
authors = [
    {name = "detection-team", email = "samuel.hudec@gmail.com"}, # fill this
]
description = "I want to see via ML" # fill this
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "torch==1.13.1",
    "pytorch-lightning==1.9.5",
    "transformers==4.25.1",
    "pandas==2.1.0",
    "scikit-learn==1.2.2",
    "numpy==1.23.5",
    "Pillow",
    "matplotlib",
]

[project.optional-dependencies]
dev = [
    "jupyter==1.*",
    "mypy == 1.2.*",
    "black == 22.*",
    "flake8 == 5.*",
    "isort==5.12.*",
    "pytest == 7.*",
    "pytest-cov == 4.*",
    "pytest-mock == 3.*"
]

[tool.black]
line-length = 120
target-version = ['py39']

[tool.isort]
known_first_party = 'ml-detection' # fill this
py_version = '39'
profile = 'black'
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120
order_by_type = false
skip = [
    '.git',
    '.github',
    '.mypy_cache',
    '.pytest_cache',
    '.venv',
    'venv',
    '.env',
    '.idea',
    'data_local',
    'venv',
    'workdir'
]

[tool.mypy]
python_version = "3.9"
disallow_untyped_defs = true
show_error_codes = true
ignore_missing_imports = true
exclude = ["venv"]